misc: add some tool installer
