+++ crates/pixi_build_frontend/Cargo.toml
@@ -46,3 +46,8 @@ pixi_build_types = { path = "../pixi_build_types" }
 insta = { workspace = true, features = ["yaml", "filters"] }
 rstest = { workspace = true }
 tempfile = { workspace = true }
+tokio = { workspace = true, features = [
+  "process",
+  "io-std",
+  "rt-multi-thread",
+] }
